move implementation to ruff_db, wire it up
this is a clean move of all the Serialize code and the two helper methods
